Merged parts of Jini's activation and reset code.
Code process for activating a user is now present. This feature is still
not complete and will require further work. Things like errors are not
very reported at the moment.

Users can now reset there password. However there is not sort of
protection around this for making sure a user is who they say there are.

The work in this commit is still not complete.